+++ datatypes-jws/src/commonMain/kotlin/at/asitplus/crypto/datatypes/jws/JsonWebKey.kt
@@ -331,8 +331,8 @@ fun CryptoPublicKey.toJsonWebKey(): JsonWebKey =
                 type = JwkType.EC,
                 keyId = jwkId,
                 curve = curve,
-                x = x,
-                y = y
+                x = xBytes,
+                y = yBytes
             )